modified:   Makefile 	modified:   ft_atoi.c 	modified:   ft_bzero.c 	modified:   ft_calloc.c 	modified:   ft_itoa.c 	new file:   ft_lstadd_back_bonus.c 	new file:   ft_lstadd_front_bonus.c 	new file:   ft_lstlast_bonus.c 	new file:   ft_lstnew_bonus.c 	new file:   ft_lstsize_bonus.c 	modified:   ft_memcpy.c 	modified:   ft_split.c 	modified:   ft_strdup.c 	modified:   ft_strjoin.c 	modified:   ft_strlcat.c 	modified:   ft_strlcpy.c 	modified:   ft_strlen.c 	modified:   ft_substr.c 	modified:   libft.h

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: ael-mous <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/02 07:44:35 by ael-mous          #+#    #+#             */
-/*   Updated: 2021/11/02 07:56:47 by ael-mous         ###   ########.fr       */
+/*   Updated: 2021/11/22 22:53:01 by ael-mous         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *c)
+#include <stddef.h>
+
+size_t	ft_strlen(const char *c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (c[i] != 0)
